change data type

+++ src/GreyScaleCalculator.cpp
@@ -103,7 +103,7 @@ double GreyScaleCalculator::CUDA_greyscale() {
     cudaTextureObject_t tex_img = 0;
     cudaCreateTextureObject(&tex_img, &resDesc, &texDesc, nullptr);
     // Setup initial value
-    uint8_t* pixel_sum = new uint8_t[1];
+    int* pixel_sum = new int[1];
     pixel_sum[0] = 0;
     cout << "pixel sum: " << pixel_sum[0] << endl;
     // Call CUDA function
